+++ 1.20/CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v8.0.1-1.20.1] - 2023-07-19
+### Changed
+- The advanced infuser can now only repair tools and armor items to prevent exploits with items from other mods that have durability
+- The config allows for reverting to the previous behavior where any damaged item can be repaired
+
 ## [v8.0.0-1.20.1] - 2023-06-27
 - Ported to Minecraft 1.20.1